05-wx_apps/wx_apps/requirements.txt
flask
flask-script
flask-sqlalchemy
flask-migrate

mysqlclient

pyyaml
